Trabajando con el envio de correo para citas

+++ BussinessLayer/DTOs/ModuloGeneral/Email/GnEmailMessageDto.cs
@@ -11,7 +11,7 @@ namespace BussinessLayer.DTOs.ModuloGeneral.Email
         public string Body { get; set; } = string.Empty;
         public bool IsHtml { get; set; }
         public List<IFormFile>? Attachments { get; set; }
-        public int EmpresaId { get; set; }
+        public long EmpresaId { get; set; }
         public int ConfiguracionId { get; set; }
     }
 }
